refactor(github-workflow): migrate services to flat SDK boundary (#10993)
Migrates 13 entities from ai/mcp/server/github-workflow/services/ to the
canonical ai/services/github-workflow/ flat SDK boundary, mirroring the
M6 sub-1 KB pattern (#10991 / PR #10995). Updates ai/services.mjs alias
imports, GH-WF Server.mjs getDependentServices override, and 5 test specs
relocated to canonical SDK test location.

Scope:
- 10 services + queries/ (7 files) + sync/ (5 files) + toolService.mjs
- All 7 currently-aliased GH_* services preserved at ai/services.mjs
  (consumer-side contract unchanged)
- Internal cross-references updated for new relative paths:
  - ../config.mjs → ../../mcp/server/github-workflow/config.mjs
  - ../logger.mjs → ../../mcp/server/github-workflow/logger.mjs
  - ../../../../../src/core/Base.mjs → ../../../src/core/Base.mjs
  - sync/ subsubdir analogous shift (one extra level deep)
- 5 unit specs moved from test/playwright/unit/ai/mcp/server/github-workflow/
  to test/playwright/unit/ai/services/github-workflow/ matching new SDK shape
- Test path math: 7→6 ../ + setup.mjs 5→4 ../
- @see refs in ai/scripts/detectTruncatedTimelines.mjs +
  refetchTruncatedIssues.mjs updated to new SDK paths
- ToolRegistration.spec.mjs path.resolve updated

Internal/non-exported: AgentStateService, DiscussionService, GraphqlService
move alongside (not currently SDK-exported but follow services/ unification
per parent epic #10986 + sibling-pattern with #10995 KB).
