+++ ai/services/github-workflow/IssueService.mjs
@@ -1,8 +1,8 @@
-import aiConfig          from '../config.mjs';
-import Base              from '../../../../../src/core/Base.mjs';
+import aiConfig          from '../../mcp/server/github-workflow/config.mjs';
+import Base              from '../../../src/core/Base.mjs';
 import GraphqlService    from './GraphqlService.mjs';
 import RepositoryService from './RepositoryService.mjs';
-import logger            from '../logger.mjs';
+import logger            from '../../mcp/server/github-workflow/logger.mjs';
 import {exec}            from 'child_process';
 import {promisify}       from 'util';
 import {spawn}           from 'child_process';
