+++ ai/services/github-workflow/SyncService.mjs
@@ -1,6 +1,6 @@
-import aiConfig        from '../config.mjs';
-import Base            from '../../../../../src/core/Base.mjs';
-import logger          from '../logger.mjs';
+import aiConfig        from '../../mcp/server/github-workflow/config.mjs';
+import Base            from '../../../src/core/Base.mjs';
+import logger          from '../../mcp/server/github-workflow/logger.mjs';
 import HealthService   from './HealthService.mjs';
 import IssueSyncer     from './sync/IssueSyncer.mjs';
 import MetadataManager from './sync/MetadataManager.mjs';
