+++ ai/services/github-workflow/sync/IssueSyncer.mjs
@@ -1,8 +1,8 @@
-import aiConfig                                      from '../../config.mjs';
-import Base                                          from '../../../../../../src/core/Base.mjs';
+import aiConfig                                      from '../../../mcp/server/github-workflow/config.mjs';
+import Base                                          from '../../../../src/core/Base.mjs';
 import crypto                                        from 'crypto';
 import fs                                            from 'fs/promises';
-import logger                                        from '../../logger.mjs';
+import logger                                        from '../../../mcp/server/github-workflow/logger.mjs';
 import matter                                        from 'gray-matter';
 import path                                          from 'path';
 import GraphqlService                                from '../GraphqlService.mjs';
