+++ test/playwright/unit/ai/services/github-workflow/DiscussionService.spec.mjs
@@ -1,4 +1,4 @@
-import {setup} from '../../../../../setup.mjs';
+import {setup} from '../../../../setup.mjs';
 
 const appName = 'DiscussionServiceTest';
 
@@ -14,9 +14,9 @@ setup({
 });
 
 import {test, expect}  from '@playwright/test';
-import Neo             from '../../../../../../../src/Neo.mjs';
-import * as core       from '../../../../../../../src/core/_export.mjs';
-import InstanceManager from '../../../../../../../src/manager/Instance.mjs';
+import Neo             from '../../../../../../src/Neo.mjs';
+import * as core       from '../../../../../../src/core/_export.mjs';
+import InstanceManager from '../../../../../../src/manager/Instance.mjs';
 
 /**
  * @summary Contract coverage for `DiscussionService.manageDiscussionComment` enriched return shape (#10841).
@@ -27,8 +27,8 @@ test.describe('Neo.ai.mcp.server.github-workflow.services.DiscussionService — 
     let originalQuery;
 
     test.beforeAll(async () => {
-        GraphqlService    = (await import('../../../../../../../ai/mcp/server/github-workflow/services/GraphqlService.mjs')).default;
-        DiscussionService = (await import('../../../../../../../ai/mcp/server/github-workflow/services/DiscussionService.mjs')).default;
+        GraphqlService    = (await import('../../../../../../ai/services/github-workflow/GraphqlService.mjs')).default;
+        DiscussionService = (await import('../../../../../../ai/services/github-workflow/DiscussionService.mjs')).default;
 
         originalQuery = GraphqlService.query.bind(GraphqlService);
     });
