+++ test/playwright/unit/ai/services/github-workflow/LabelService.spec.mjs
@@ -1,4 +1,4 @@
-import {setup} from '../../../../../setup.mjs';
+import {setup} from '../../../../setup.mjs';
 
 const appName = 'LabelServiceTest';
 
@@ -14,9 +14,9 @@ setup({
 });
 
 import {test, expect}  from '@playwright/test';
-import Neo             from '../../../../../../../src/Neo.mjs';
-import * as core       from '../../../../../../../src/core/_export.mjs';
-import InstanceManager from '../../../../../../../src/manager/Instance.mjs';
+import Neo             from '../../../../../../src/Neo.mjs';
+import * as core       from '../../../../../../src/core/_export.mjs';
+import InstanceManager from '../../../../../../src/manager/Instance.mjs';
 
 /**
  * @summary Contract coverage for `LabelService.listLabels` error propagation (#10112).
@@ -38,8 +38,8 @@ test.describe('Neo.ai.mcp.server.github-workflow.services.LabelService', () => {
     let originalQuery;
 
     test.beforeAll(async () => {
-        GraphqlService = (await import('../../../../../../../ai/mcp/server/github-workflow/services/GraphqlService.mjs')).default;
-        LabelService   = (await import('../../../../../../../ai/mcp/server/github-workflow/services/LabelService.mjs')).default;
+        GraphqlService = (await import('../../../../../../ai/services/github-workflow/GraphqlService.mjs')).default;
+        LabelService   = (await import('../../../../../../ai/services/github-workflow/LabelService.mjs')).default;
 
         originalQuery = GraphqlService.query.bind(GraphqlService);
     });
